+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ### Next
+  - Added support for Humidity sensor module.
   - Updated Google Play Services to 7.8.0
   - Added support for Android Studio 1.3.1
   - Updated gradle to 1.3.0